+++ src/users/users.controller.ts
@@ -1,15 +1,5 @@
-import {
-  Body,
-  Controller,
-  Get,
-  Param,
-  ParseIntPipe,
-  Post,
-  Request,
-} from '@nestjs/common';
+import { Controller, Get, Param, ParseIntPipe, Request } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { Public } from '../decorators/public.decorator';
-import { CreateUserRequestDto } from './dto/create-user-request.dto';
 
 @Controller('users')
 export class UsersController {
